Moved details screens to their own directory
